Implement checkmate practice for the redesign
Turns the /checkmatepractice stub into a working page. The board and its side
bar occupy the page; the checkmate list is an overlay above it that hides on
selection and reopens from the side bar. Games are local — nothing transmitted,
no server game created.

The selection list is SSR'd: practicePageController resolves each checkmate id
to its piece icons, so the client no longer fetches piece svgs or regenerates
them on theme change. pieceSvgCache now keeps white's and black's variants
rather than black alone, since the artwork distinguishes them.

gamesession gains a 'practice' session type. Registering these games as
'online' (as the ported code did) would have sent a socket resign intent when
the engine failed to load, and waited on a server that isn't there to conclude
the game. enginegame reads its role off gamesession instead of the game page's
gamePageData, which practice pages don't define.

Practice positions now play inside a world border, resolving the tripwire left
on startEngineGame. Side bar undo/restart buttons replace the dead
guigameinfo-* document events.
